Added working button to change box size
When the button is clicked, a prompt asks a number from the user.
If the result is not a number, or not between 1 and 100, the prompt loops.
With this new number, we empty our container first with another while loops
(in this loop, so long as container has a child, it will be removed until
none are left). Then, we use the result from the prompt to call our previous
etch function and create and new box with the prompted size and the event
listener.